Added:
Takes a chessboard size as input and prints the maximum number of bishops that can be arranged on the chessboard in such a way that no bishop attacks any
other bishop, and prints the chessboard states in how many ways this can be done.

Code that takes input for:
1. Number of columns: m
2. Number of rows: n

Output:

Displays number of chessboards that have maximum number of bishops arranged is such a way that they don't attack each others.